# convert void 0 to a new var