I have sprinkled the code with lots of TODOs and questions. The basic implementation works, using only within-same-function transforms as proposed by @kzc.

- According to the benchmark:
     - Average uglified size improved by ~0.16%
     - Average gzip size worsened by ~0.17%

Do these results mean that we should abandon this idea? If we want this transform, it shouldn't be enabled by default, and information needs to be added to the README.

<details>
<summary>Results from node test/benchmark.js</summary>

<table><tr>

<td>
BEFORE THIS CHANGE

https://code.jquery.com/jquery-3.2.1.js
- parse: 1.107s
- rename: 0.769s
- compress: 8.914s
- scope: 0.274s
- mangle: 0.640s
- properties: 0.000s
- output: 0.539s
- total: 12.243s

Original: 268039 bytes
Uglified: 86074 bytes
GZipped:  30058 bytes
SHA1 sum: 3e16e5ec39f60f779cf1cb407dcc0fb93c1216ab

https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.4/angular.js
- parse: 2.029s
- rename: 1.359s
- compress: 15.630s
- scope: 0.598s
- mangle: 1.171s
- properties: 0.000s
- output: 0.562s
- total: 21.349s

Original: 1249863 bytes
Uglified: 173616 bytes
GZipped:  59988 bytes
SHA1 sum: 87f388ccba07cb8505ac03ff75cadb235c6a437a

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
- parse: 4.141s
- rename: 2.746s
- compress: 23.607s
- scope: 0.304s
- mangle: 0.559s
- properties: 0.000s
- output: 0.430s
- total: 31.788s

Original: 1590107 bytes
Uglified: 466740 bytes
GZipped:  118842 bytes
SHA1 sum: f8502b2990e245757a06d53ad670aaa710806e66

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
- parse: 0.524s
- rename: 0.373s
- compress: 3.588s
- scope: 0.167s
- mangle: 0.283s
- properties: 0.000s
- output: 0.222s
- total: 5.157s

Original: 69707 bytes
Uglified: 36836 bytes
GZipped:  9582 bytes
SHA1 sum: 31b3a3cb53ed8dd5e92519891b92a60eb07babfc

https://unpkg.com/react@15.3.2/dist/react.js
- parse: 1.962s
- rename: 1.611s
- compress: 14.967s
- scope: 0.407s
- mangle: 0.788s
- properties: 0.000s
- output: 0.501s
- total: 20.237s

Original: 701412 bytes
Uglified: 205174 bytes
GZipped:  62273 bytes
SHA1 sum: 7a8ad6431a54c84e476a1a87aeba0a80a6f1dd06

http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
- parse: 4.303s
- rename: 3.194s
- compress: 19.328s
- scope: 0.374s
- mangle: 0.676s
- properties: 0.000s
- output: 0.790s
- total: 28.665s

Original: 1852178 bytes
Uglified: 526064 bytes
GZipped:  129050 bytes
SHA1 sum: 22037ab14a33379b2ee9ba9270ce22e4ee970a2c

https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
- parse: 1.183s
- rename: 0.735s
- compress: 14.983s
- scope: 0.286s
- mangle: 0.534s
- properties: 0.000s
- output: 0.401s
- total: 18.122s

Original: 539590 bytes
Uglified: 69578 bytes
GZipped:  24877 bytes
SHA1 sum: 025d83a7526aa44dd4a175a5fabc3ef7c6992967

https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
- parse: 2.478s
- rename: 2.393s
- compress: 18.957s
- scope: 0.398s
- mangle: 0.854s
- properties: 0.000s
- output: 0.681s
- total: 25.761s

Original: 451131 bytes
Uglified: 211601 bytes
GZipped:  71051 bytes
SHA1 sum: 1636ad62c5338200bb99eb57fa3fd39f882502d2

https://raw.githubusercontent.com/kangax/html-minifier/v3.5.7/dist/htmlminifier.js
- parse: 3.678s
- rename: 3.485s
- compress: 22.196s
- scope: 0.302s
- mangle: 0.647s
- properties: 0.000s
- output: 0.489s
- total: 30.797s

Original: 1063075 bytes
Uglified: 508164 bytes
GZipped:  158013 bytes
SHA1 sum: ad6e5c94ba84dc6cb0227b57504f4c2e3abac535
</td>

<td>
WITH THIS CHANGE

https://code.jquery.com/jquery-3.2.1.js
- parse: 1.204s
- rename: 0.644s
- compress: 8.808s
- scope: 0.388s
- mangle: 0.740s
- properties: 0.000s
- output: 0.478s
- total: 12.263s

Original: 268039 bytes
Uglified: 85768 bytes
GZipped:  30113 bytes
SHA1 sum: 5a7980c8fc4be9cda00bdf23457b336d728021ff

https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.4/angular.js
- parse: 1.966s
- rename: 1.255s
- compress: 16.213s
- scope: 0.789s
- mangle: 1.462s
- properties: 0.000s
- output: 0.928s
- total: 22.613s

Original: 1249863 bytes
Uglified: 173393 bytes
GZipped:  60045 bytes
SHA1 sum: 6cc222e3f8bc88192bba299f3504ccc3959fa940

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
- parse: 4.036s
- rename: 2.125s
- compress: 26.278s
- scope: 0.318s
- mangle: 0.650s
- properties: 0.000s
- output: 0.508s
- total: 33.916s

Original: 1590107 bytes
Uglified: 466184 bytes
GZipped:  119122 bytes
SHA1 sum: 660b9e954b1f85ba742e964bbb35ebad167664e3

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
- parse: 0.625s
- rename: 0.336s
- compress: 2.977s
- scope: 0.105s
- mangle: 0.279s
- properties: 0.000s
- output: 0.224s
- total: 4.546s

Original: 69707 bytes
Uglified: 36830 bytes
GZipped:  9580 bytes
SHA1 sum: d6c3dab88ac5f13c1f5dd20c232b60e3efbe237b

https://unpkg.com/react@15.3.2/dist/react.js
- parse: 2.177s
- rename: 1.265s
- compress: 15.991s
- scope: 0.350s
- mangle: 1.673s
- properties: 0.000s
- output: 0.803s
- total: 22.259s

Original: 701412 bytes
Uglified: 204937 bytes
GZipped:  62343 bytes
SHA1 sum: a76418a746a313e3eab3db474811dbf280758922

http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
- parse: 3.680s
- rename: 3.006s
- compress: 21.636s
- scope: 0.365s
- mangle: 0.745s
- properties: 0.000s
- output: 0.481s
- total: 29.913s

Original: 1852178 bytes
Uglified: 524078 bytes
GZipped:  129669 bytes
SHA1 sum: 1f19d7f41160f151fca54afe903b0a34d3098f1c

https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
- parse: 1.173s
- rename: 0.805s
- compress: 14.316s
- scope: 0.243s
- mangle: 0.665s
- properties: 0.000s
- output: 0.444s
- total: 17.646s

Original: 539590 bytes
Uglified: 69578 bytes
GZipped:  24877 bytes
SHA1 sum: 025d83a7526aa44dd4a175a5fabc3ef7c6992967

https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
- parse: 2.650s
- rename: 2.019s
- compress: 21.041s
- scope: 0.415s
- mangle: 0.923s
- properties: 0.000s
- output: 0.583s
- total: 27.631s

Original: 451131 bytes
Uglified: 211569 bytes
GZipped:  71062 bytes
SHA1 sum: 5c47a60b136f38376fcadff4129f0254ebca0d5e

https://raw.githubusercontent.com/kangax/html-minifier/v3.5.7/dist/htmlminifier.js
- parse: 3.437s
- rename: 2.979s
- compress: 24.390s
- scope: 0.312s
- mangle: 0.657s
- properties: 0.000s
- output: 0.532s
- total: 32.307s

Original: 1063075 bytes
Uglified: 507867 bytes
GZipped:  158078 bytes
SHA1 sum: 87d89bfcd4d713ed3e91f9bfecd7dc86c59cd00b

</td>

</tr></table>

</details>


fixes #2585