@alexlamsl 2018-02-12
The uglified numbers look positive for starters.

I suspect the worse gzipped results are (partly) due to `mangle` – one way to assess is to run without it:
```
node test/benchnark.js -c
```

I haven't looked at the code yet as I'm about to doze off – will get back to you on that.

Thanks for working on this :+1: