@Skalman 2018-02-12
I *think* disabling mangling doesn't change the gzip results (~0.17% worse, like before). I'm also tired, so I might be mistaken.

Here's an idea that probably doesn't change anything, but *could*.
```
// with this transform
var existing=1,variables=2,undefined; ...

// idea
var undefined,existing=1,variables=2;...
```