@alexlamsl 2018-02-13
Re-reading https://github.com/mishoo/UglifyJS2/issues/2585#issuecomment-351479601 it claims to produce **post-gzip** improvements, which is not observed anywhere (except `bootstrap.js`) - so I'm not sure how was done.