@alexlamsl 2018-02-13
Here's a more restricted version which never introduces new variables, i.e. only group `void 0`s into an existing variable which happens to be `undefined`:
```diff
--- a/lib/compress.js
+++ b/lib/compress.js
@@ -58,6 +58,7 @@ function Compressor(options, false_by_default) {
         evaluate      : !false_by_default,
         expression    : false,
         global_defs   : {},
+        group_voids   : !false_by_default,
         hoist_funs    : false,
         hoist_props   : !false_by_default,
         hoist_vars    : false,
@@ -196,6 +197,11 @@ merge(Compressor.prototype, {
                 }
             }
         }
+        if (this.option("group_voids")) {
+            node.figure_out_scope(mangle);
+            node.reset_opt_flags(this);
+            node.group_voids(this);
+        }
         if (this.option("expression")) {
             node.process_expression(false);
         }
@@ -270,6 +276,28 @@ merge(Compressor.prototype, {
         return this.TYPE == node.TYPE && this.print_to_string() == node.print_to_string();
     });

+    AST_Toplevel.DEFMETHOD("group_voids", function(compressor) {
+        var void_sym;
+        this.transform(new TreeTransformer(function(node, descend) {
+            if (node instanceof AST_Scope) {
+                var old_sym = void_sym;
+                node.variables.each(function(def) {
+                    if (old_sym !== void_sym) return;
+                    var fixed = def.orig[0].fixed_value();
+                    if (fixed && is_undefined(fixed, compressor)) {
+                        void_sym = def.orig[0];
+                    }
+                });
+                descend(node, this);
+                void_sym = old_sym;
+                return node;
+            }
+            if (void_sym && is_undefined(node, compressor)) {
+                return make_node(AST_SymbolRef, void_sym, void_sym);
+            }
+        }));
+    });
+
     AST_Scope.DEFMETHOD("process_expression", function(insert, compressor) {
         var self = this;
         var tt = new TreeTransformer(function(node) {
```
This only makes a difference to `math.js`, but even so `gzip` still isn't a win:

#### `node test/benchmark -mc group_voids=false`
```
https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
- parse: 0.875s
- rename: 0.562s
- compress: 5.350s
- scope: 0.203s
- mangle: 0.328s
- properties: 0.000s
- output: 0.234s
- total: 7.552s

Original: 1590107 bytes
Uglified: 466736 bytes
GZipped:  118836 bytes
SHA1 sum: 39105dfb8a9231469b83993c388735b9655cab2d
```

#### `node test/benchmark -mc group_voids=true`
```
https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
- parse: 0.921s
- rename: 0.579s
- compress: 5.531s
- scope: 0.156s
- mangle: 0.328s
- properties: 0.000s
- output: 0.235s
- total: 7.750s

Original: 1590107 bytes
Uglified: 466571 bytes
GZipped:  118888 bytes
SHA1 sum: 6a6435c980dea291bd5b07bd1a34c131b9d19a38
```

So I reckon the initial analysis is flawed, which is sadly quite often the case when it comes to `mangle` and `gzip`.

At least it gives you an idea how scan and replace the AST with minimal number of rounds :wink: