@Skalman 2018-02-16
My idea to place the `undefined` variable first in the `var` declaration produces slightly worse results. I've tried a few variants to see if the placement of the new `undefined` declaration can make the gzip gods happier.

```js
// Last in the first var (my initial implementation)
// Gzip worse by 0.174%
var a, b, undefined; var c, d; var e, f;

// First in the first var
// Gzip worse by 0.189%
var undefined, a, b; var c, d; var e, f;

// Last in the last var
// Gzip worse by 0.145%
var a, b; var c, d; var e, f, undefined;

// First in the last var
// Gzip worse by 0.158%
var a, b; var c, d; var undefined, e, f;
```

@alexlamsl's idea of never introducing new variables has a tiny effect compared to introducing new variables, so I don't think it would be worth it. If we're implementing any of these, the "last in the last var" should be used, since it produces the least bad gzip.

I also think we should benchmark introducing a `var` before deciding what to do.