@Skalman 2018-02-16
Not creating a new `var` (like my initial impementation) improves uglification by 0.159% (as reported in my initial comment).

```js
// Last in the last var (best from my comment above)
// Uglify better by 0.159%
// Gzip worse by 0.145%


// Last in the last var + create new `var` last in function if there is no existing `var`
// Uglify better by 0.165%
// Gzip worse by 0.162%


// Last in the last var + create new `var` last in function if there is no existing `var` only if there are more than two `void 0` uses
// Uglify better by 0.171%
// Gzip worse by 0.152%


// Last in the last var + create new `var` first in function if there is no existing `var`
// Uglify better by 0.164%
// Gzip worse by 0.152%


// Last in the last var + create new `var` first in function if there is no existing `var` only if there are more than two `void 0` uses
// Uglify better by 0.171%
// Gzip worse by 0.148%
```

Idea: introduce new `var` if there are more than three `void 0` uses. 😄