@alexlamsl 2018-02-16
Introducing new variables tend to cause cascading effects on variable name assignments within `mangle`, which is why such a large gzip delta can be observed with a seemingly small change.

I haven't come up with any better ideas to try yet, but then we are in no great hurry either :wink: