@alexlamsl 2018-02-16
And yes, putting that new variable first in the first `var` would shift the names of all subsequent declarations in the scope, so has a bigger effect than a trailing declaration (still not zero since it would cause intra-scope disruptions).