@alexlamsl 2018-02-16
FWIW, `inline>1` also had to fight with `gzip` regressions back then via various tricks and hacks to make it profitable.